Extract all helper methods related to donations into a DonationsHelper.
