+++ app/helpers/dashboard_helper.rb
@@ -12,22 +12,6 @@ module DashboardHelper
     number_with_delimiter(total || "-1")
   end
 
-  def total_received_money_donations(range = selected_range)
-    current_organization.donations.during(range).sum { |d| d.money_raised || 0 }
-  end
-
-  def total_received_money_donations_from_product_drives(range: selected_range)
-    current_organization.donations.by_source(:product_drive).during(range).sum { |d| d.money_raised || 0 }
-  end
-
-  def total_received_donations(range = selected_range)
-    number_with_delimiter total_received_donations_unformatted(range)
-  end
-
-  def total_received_from_product_drives(range = selected_range)
-    number_with_delimiter total_received_from_product_drives_unformatted(range)
-  end
-
   def total_purchased(range = selected_range)
     number_with_delimiter total_purchased_unformatted(range)
   end
@@ -45,14 +29,6 @@ module DashboardHelper
   end
 
   private
-
-  def total_received_donations_unformatted(range = selected_range)
-    LineItem.active.where(itemizable: current_organization.donations.during(range)).sum(:quantity)
-  end
-
-  def total_received_from_product_drives_unformatted(range = selected_range)
-    LineItem.active.where(itemizable: current_organization.donations.by_source(:product_drive).during(range)).sum(:quantity)
-  end
 
   def total_purchased_unformatted(range = selected_range)
     LineItem.active.where(itemizable: current_organization.purchases.during(range)).sum(:quantity)
